+++ modules/DOMelements.js
@@ -1,0 +1,4 @@
+// Getting DOM elements
+export const form = document.querySelector('form');
+export const input = document.querySelector('input');
+export const list = document.querySelector('ul');
